RuoYi-Vue3-FastAPI/ruoyi-fastapi-backend/requirements.txt
aiofiles==25.1.0
alembic==1.16.5
APScheduler==3.11.1
async-lru==2.0.5
asyncmy==0.2.10
bcrypt==5.0.0
fastapi[all]==0.125.0
loguru==0.7.3
openpyxl==3.1.5
xlrd<2.0.0
pandas==2.3.3
Pillow==11.3.0
qrcode[pil]>=7.0
psutil==7.1.3
pydantic-validation-decorator==0.1.5
PyJWT[crypto]==2.10.1
PyMySQL==1.1.2
redis==6.4.0
ruff==0.14.10
SQLAlchemy[asyncio]==2.0.45
sqlglot[rs]==28.5.0
user-agents==2.2.0

# AI模型支持
openai>=1.0.0
anthropic>=0.18.0
aiohttp>=3.9.0

# Word文档处理
python-docx>=1.1.0
